+++ HISTORY.md
@@ -1,4 +1,9 @@
+0.0.3 / 24.11.2016
+===================
+
+  * Added `format` function.
+
 0.0.1 / 24.11.2016
 ===================
 
-  * Initial release. `parse`function is working.
+  * Initial release. `parse` function is working.